requirements.txt
streamlit
requests
python-dotenv
altair